requirements.txt
torch
numpy
pandas
scipy

deepxde

mlflow
pyyaml

plotly